Use kill (not ps) to find existing ssh-agent
`ps` and the like are not entirely reliable for finding the running
ssh-agent.  For example, if procfs on Linux is mounted with hidepid > 0,
then ps and pgrep appear to fail to find a running ssh-agent under the
current user.  Sending SIGCONT (continue) should be benign and should
only succeed if the current user is able to signal the pid (either USER
owns SSH_AGENT_PID process or USER is root).  It's possible ssh-agent
with pid SSH_AGENT_PID died and left SSH_AUTH_SOCK behind and a new
receives SSH_AGENT_PID and would successfully receive SIGCONT, in which
case we'll just let the user resolve this issue manually since it should
be rare and rather unlikely.